requirements.txt
pymongo
ipykernel
pandas 
numpy
scikit-learn
python-dotenv

-e . 